handle client and change api doc
