+++ server/README.md
@@ -1,11 +1,77 @@
 # Fancy Todo Server
 This app has : 
-* RESTful endpoint for asset's CRUD operation
+* RESTful endpoint for todo's CRUD operation
 * JSON formatted response
 
 &nbsp;
 
 ## RESTful endpoints
+### POST http://localhost:3000/register
+
+> Post User Register
+
+_Request Body_
+```
+{
+    "email": "spiderman",
+    "username": "spiderman@gmail.com",
+    "password": "123456"
+}
+```
+
+_Response (201)_
+```
+{
+    "id": "1",
+    "email": "spiderman@gmail.com"
+}
+```
+
+_Response (400)_
+```
+{
+    "message": [
+        "Username must be filled",
+        "Username minimum 6 characters",
+        "Username must be unique",
+        "Email must be filled",
+        "Email not valid",
+        "Email must be unique",
+        "Password must be filled",
+        "Password at least 6 characters"
+    ]
+}
+```
+
+### POST http://localhost:3000/login
+
+> Post User Login
+
+_Request Body_
+```
+{
+    "email": "spiderman@gmail.com",
+    "password": "123456"
+}
+```
+
+_Response (200)_
+```
+{
+    "access_token": "<generated by system>",
+    "username": "spiderman"
+}
+```
+
+_Response (400)_
+```
+{
+    "message": [
+        "Email not found, please register first",
+        "Invalid Email/password"
+    ]
+}
+```
 ### GET http://localhost:3000/todos
 
 > Get all todo list
@@ -26,6 +92,9 @@ _Response (200 - OK)_
     "due_date": "2021-01-05T00:00:00.149Z",
     "createdAt": "2021-01-04T07:15:12.149Z",
     "updatedAt": "2021-01-04T07:15:12.149Z",
+    "User": {
+      "username": "superman"
+    }
   },
   {
     "id": 2,
@@ -35,6 +104,9 @@ _Response (200 - OK)_
     "due_date": "2021-01-05T00:00:00.149Z",
     "createdAt": "2021-01-04T07:15:12.149Z",
     "updatedAt": "2021-01-04T07:15:12.149Z",
+    "User": {
+      "username": "spiderman"
+    }
   }
 ]
 ```
