+++ server/controllers/apiController.js
@@ -1,23 +1,16 @@
 const axios = require('axios')
 
-class ApiController{
-    static getNews(req, res, next){
-        const api_key = process.env.API_KEY
-        const url = `http://newsapi.org/v2/top-headlines?country=id&category=technology&apiKey=${api_key}`
-        axios.get(url)
-        .then(response => {
-            res.status(200).json(response.data)
+class ApiController {
+    static getWeather(req, res, next) {
+        const apiKey = process.env.WEATHER_API_KEY
+        const unit = "metric";
+        const url = `https://api.openweathermap.org/data/2.5/weather?q=jakarta&appid=${apiKey}&units=${unit}`;
+        axios({
+            method: 'GET',
+            url: url,
         })
-        .catch(err => {
-            next(err)
-        })
-    }
-
-    static getQuotes(req, res, next){
-        axios.get('https://type.fit/api/quotes')
-        .then(response => {
-            // console.log(response.data);
-            res.status(200).json(response.data)
+        .then(result => {
+            res.status(200).json(result.data)
         })
         .catch(err => {
             next(err)
