+++ server/routes/todoRoutes.js
@@ -3,8 +3,7 @@ const { authorized } = require('../middlewares')
 const express = require('express')
 const router = express.Router()
 
-router.get('/news', ApiController.getNews)
-router.get('/quotes', ApiController.getQuotes)
+router.get('/weather', ApiController.getWeather)
 router.get('/', TodoController.getTodos)
 router.post('/', TodoController.addTodo)
 router.get('/:id', authorized, TodoController.getTodoId)
